requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Session==0.5.0
gunicorn==21.2.0
gevent==23.7.0
greenlet==3.0.2
python-dotenv==1.0.0

# Database
groq==0.4.1
pymongo==4.6.1

# Authentication and Security
bcrypt==4.1.2

# Real-time Communication
python-socketio==5.10.0
python-engineio==4.9.0

# Logging
structlog==24.2.0

# Email
email-validator==2.1.0.post1
